Add files via upload
This zip file contains the AuoWheel java project, can be opened in any java based IDE. This project is for the simulation of path creation and shows how the algorithm runs. The code is written, for the purpose of importing on raspberry pi 3. It contains various packages which show the modules of the system.